Don’t print out module info when ignored
There was a bug that outputted misleading summary & detailed info. If --ignore-module was used, it would ignore it in the results calculation (correct) but be misleading and still say “COVERED” or 100% in the output.

As well, if a file was empty, and --ignore-module was used, it would still show the module in the output.